Single source of validation & update SearchBar icons after error is resolved
